requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
packaging==21.3
Pillow==9.2.0
playsound==1.3.0
pyparsing==3.0.9
pyserial==3.5
redis==4.3.4
requests==2.28.1
urllib3==1.26.12
wrapt==1.14.1
